refactor(services): extract 24h cache wrapper to community-cache.ts (B-025)
Second slice. The `loadCommunityPackagesWithCache` function +
`PluginWithApp` interface (renamed to `PluginCacheHost` for
narrower scope) move to a dedicated `community-cache.ts`.

The new file is responsible for:
  - 24-hour cache TTL (constant lifted to module-level, was
    inline magic number)
  - Settings-backed cache storage / persistence
  - Stale-cache fallback when the live load fails
  - Empty-result caching (avoids hammering 404'd directories)

`community-packages.ts` re-exports `loadCommunityPackagesWithCache`
+ `PluginCacheHost` so existing callers (`PackageBrowser.ts`,
`src/types.ts`) don't see the move — keeps this PR's diff inside
the services/ tree.

`loadAllCommunityPackages` router now takes `PluginCacheHost`
instead of the old `PluginWithApp`. Functionally identical (same
shape, narrower contract), but the new name reflects intent:
this is the cache's host, not "anything with an app reference".

350/350 tests still pass.
